Replace illegal characters from each value as the new track filename is built
Previously, the illegal characters were only replaced after the entire string was built. This meant that if the track name/album/artist had a slash in it, that was not replaced because slashes were permitted. Now, slashes are not permitted and are replaced as the track is built, which solves this problem.